Write tests in TypeScript
