+++ jest.config.mjs
@@ -115,7 +115,7 @@ export default {
   // restoreMocks: false,
 
   // The root directory that Jest should scan for tests and modules within
-  // rootDir: undefined,
+  rootDir: "./dist",
 
   // A list of paths to directories that Jest should use to search for files in
   // roots: [
