--- test/sanity.test.js
@@ -1,7 +1,0 @@
-describe("Sanity Check", () => {
-  test("true should be true", () => {
-    expect(true).toBe(true);
-  });
-
-  // eof
-});
