+++ test/sanity.test.ts
@@ -1,0 +1,7 @@
+describe("Sanity Check", (): void => {
+  test("true should be true", (): void => {
+    expect(true).toBe(true);
+  });
+
+  // eof
+});
