--- test/setup-teardown.js
@@ -1,9 +1,0 @@
-const teardown = async () => {
-  console.log("TESTSUITE TEARDOWN");
-  if (global.fileSentinel) {
-    await global.fileSentinel.terminate();
-  }
-  console.log("TESTSUITE TEARDOWN DONE");
-};
-
-export default teardown;
